pyproject.toml
[tool.poetry]
name = "django-shop"
version = "0.1.0"
description = ""
authors = ["yetian29 <dathien000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-ninja = "^1.3.0"
django-environ = "^0.11.2"
punq = "^0.7.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs,
    "F401"
]